Selectors: Update isPrivateSite selector to fallback to settings state
The isPrivateSite selector now falls back to site settings state if site state
isn't available. This commit updates the unit tests to match this behavior.

With the updated behavior in isPrivateSite, tests for
isSiteSupportingImageEditor() were failing. Updated the first test to include a
blank siteSettings object so tests would pass.